\name{as.CompoundFactor}
\alias{as.CompoundFactor}
\alias{as.CompoundFactor.CompoundFactor}
\alias{as.CompoundFactor.SpotRate}
\title{as.CompoundFactor}
\usage{
  as.CompoundFactor(object, ...)

  \method{as.CompoundFactor}{SpotRate} (object)

  \method{as.CompoundFactor}{CompoundFactor} (object)
}
\description{
  Compound factor
}

