% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardrate-class.R
\name{as.forwardrate-method}
\alias{as.forwardrate-method}
\alias{as.forwardrate}
\alias{as.forwardrate.SpotRate}
\alias{as.forwardrate.SpotRateCurve}
\title{Coerce objects to ForwardRate}
\usage{
\method{as.forwardrate}{SpotRate}(x, terms, ...)

\method{as.forwardrate}{SpotRateCurve}(x, ...)
}
\arguments{
\item{x}{a SpotRate or a SpotRateCurve object.}

\item{terms}{a numeric with positive values representing terms or a Term
object.}

\item{...}{additional arguments}
}
\description{
A \code{ForwardRate} object can be created from a SpotRate object and
a SpotRateCurve.
}
