\name{as.SpotRate}
\alias{as.SpotRate}
\alias{as.SpotRate.CompoundFactor}
\alias{as.SpotRate.SpotRate}
\title{as.SpotRate}
\usage{
  as.SpotRate(object, ...)

  \method{as.SpotRate}{SpotRate} (object,
    term = term(object))

  \method{as.SpotRate}{CompoundFactor} (object, dib = 252,
    compounding = "compounded")
}
\description{
  Coerces to a SpotRate
}

