% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotratecurve-class.R
\name{as.spotratecurve-method}
\alias{as.spotratecurve-method}
\alias{as.spotratecurve}
\alias{as.spotratecurve.ForwardRate}
\title{Coerce objects to spotratecurve}
\usage{
as.spotratecurve(x, ...)

\method{as.spotratecurve}{ForwardRate}(x, refdate = Sys.Date(), ...)
}
\arguments{
\item{x}{a ForwardRate object.}

\item{...}{additional arguments}

\item{refdate}{the curve reference date.}
}
\value{
a SpotRateCurve object
}
\description{
A SpotRateCurve can be created from a \code{ForwardRate} object.
}
