\name{as.spotratecurve}
\alias{as.spotratecurve}
\title{Coerce objects to spotratecurve}
\usage{
as.spotratecurve(obj, ...)
}
\arguments{
  \item{object}{should be a \code{matrix} or a
  \code{data.frame}}

  \item{...}{other spotratecurve's parameters}
}
\value{
spotratecurve object
}
\description{
A spotratecurve can be coerced from \code{matrix} and
\code{data.frame}.
}

