% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-class.R
\name{as.term}
\alias{as.term}
\alias{as.term,character-method}
\title{Coerce a character to a Term}
\usage{
as.term(x, ...)
}
\arguments{
\item{x}{a character to be coerced to a Term.}

\item{...}{additional arguments. Currently unused.}
}
\value{
A \code{Term} object created from a string.
}
\description{
\code{as.term} coerces a character vector to a Term object.
}
\details{
The string representation of the Term class follows the layout:

\preformatted{NUMBER UNITS}

where units is one of: days, months, years.
}
\examples{
t <- as.term("6 months")
}
