\name{compound.factor}
\alias{compound.factor}
\alias{compound.factor.CompoundFactor}
\alias{compound.factor.SpotRate}
\title{compound.factor}
\usage{
compound.factor(object, ...)

\method{compound.factor}{CompoundFactor}(object)

\method{compound.factor}{SpotRate}(object)
}
\description{
Compound factor
}

