\name{compound}
\alias{compound}
\alias{compound.spotrate}
\title{Compound function}
\usage{
compound(obj, ...)

\method{compound}{spotrate}(obj, term, units = NULL, from = NULL,
  to = NULL, ...)
}
\arguments{
  \item{obj}{.}

  \item{...}{.}
}
\description{
Computes the compounding factor
}
\details{
Compounding factor
}

