\name{compounding}
\alias{compounding}
\alias{compounding.spotrate}
\title{Compounding accessor method}
\usage{
compounding(obj, ...)

\method{compounding}{spotrate}(obj, ...)
}
\arguments{
  \item{obj}{any object}

  \item{...}{unused extra arguments}
}
\description{
Compounding method
}
\details{
Compounding method
}

