\name{CurveInterpolation}
\alias{CurveInterpolation}
\title{CurveInterpolation}
\usage{
  CurveInterpolation(curve, method = "flatforward")
}
\arguments{
  \item{curve}{curve}

  \item{method}{\code{flatforward}, \code{linear},
  \code{loglinear}, \code{spline}, \code{hermite},
  \code{monotone}}
}
\value{
  \code{CurveInterpolation} class
}
\description{
  Creates a curve interpolation class
}

