% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daycount-class.R
\docType{class}
\name{Daycount-class}
\alias{Daycount-class}
\title{Daycount class}
\description{
\code{Daycount} class helps adjusting the terms to compound interest rates.
With annual rates it is necessary to convert periods of days or months
to years units.
The day count convention helps with that by defining the number of days of
one year.
Together with a calendar it defines the way the wordays are counted between
two dates.
}
\details{
Common day count rules are: \code{actual/365}, \code{actual/360},
\code{business/252}, \code{30/360}, ...
}
