\name{daycount}
\alias{daycount}
\alias{daycount.spotrate}
\title{Day count accessor}
\usage{
daycount(obj, ...)

\method{daycount}{spotrate}(obj, ...)
}
\arguments{
  \item{obj}{any object which has a daycount attribute}

  \item{...}{extra arguments}
}
\value{
a \code{daycount} object
}
\description{
Method to access the \code{daycount} attribute.
}

