% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daycount-class.R
\name{dib}
\alias{dib}
\alias{dib,Daycount-method}
\title{Days in base for Daycount}
\usage{
dib(x)
}
\arguments{
\item{x}{a Daycount object.}
}
\value{
A numeric.
}
\description{
\code{dib} returns the days in base, that is the number of days used to
define one year.
}
\details{
The method \code{dib} returns the days in base for a daycount convention.
Since we work with annual rates the days in base define the amount of days
in a year used in the convention.
}
\examples{
dc <- daycount("actual/360")
dib(dc)
}
