% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-class.R
\name{diff,Term-method}
\alias{diff,Term-method}
\title{Calculate lagged differences of Term objects}
\usage{
\S4method{diff}{Term}(x, lag = 1, ..., fill = NULL)
}
\arguments{
\item{x}{a Term object.}

\item{lag}{a numerix indicating which lag to use.}

\item{...}{additional arguments. Currently unused.}

\item{fill}{a numeric value (or \code{NA}) to fill the empty created by
applying diff to a Term object.}
}
\value{
A new \code{Term} object with lagged differences of the given \code{Term} object.
}
\description{
\code{diff} returns a Term vector with lagged differences.
}
\examples{
t <- term(1:10, "months")
diff(t)
}
