\docType{package}
\name{fixedincome-package}
\alias{fixedincome}
\alias{fixedincome-package}
\title{Tools for fixed income calculations}
\description{
Tools for calculations on fixed income, term structure and
interest rate.
}
\details{
\code{fixedincome} brings a rich set of tools to make fixed
income calculations fairly easy. Fixed income calculations
have many specific rules which deal with subjects ranging
from compounding regimes to business days rules to compute
the amount of non-working days between maturities.
}
\author{
Wilson Freitas \email{wilson.freitas@gmail.com}
}
\references{
Frank Fabozzi. Fixed Income Mathematics, Wiley, 1994.

Bruce Tuckman. Fixed Income Securities, Wiley, 1994.
}

