\name{forward.rate}
\alias{forward.rate}
\alias{forward.rate.SpotRate}
\alias{forward.rate.SpotRateCurve}
\title{forward.rate}
\usage{
forward.rate(object, ...)

\method{forward.rate}{SpotRate}(object, other)

\method{forward.rate}{SpotRateCurve}(curve, from.term, to.term = NULL,
  forward.term = NULL)
}
\description{
Forward rate
}

