% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardrate-class.R
\name{forwardrate-method}
\alias{forwardrate-method}
\alias{forwardrate}
\alias{forwardrate.numeric}
\alias{forwardrate.SpotRateCurve}
\title{Create a SpotRateCurve object}
\usage{
\method{forwardrate}{numeric}(x, terms, compounding, daycount, calendar, .copyfrom = NULL, ...)

\method{forwardrate}{SpotRateCurve}(x, t1 = NULL, t2 = NULL, ...)
}
\arguments{
\item{terms}{a numeric vector with positive values representing terms of
the forward rates.}

\item{compounding}{a character with the compouning name.}

\item{daycount}{a character representing the daycount.}

\item{calendar}{a calendar object.}

\item{.copyfrom}{a SpotRate object that is used as reference to build
the SpotRateCurve object.}

\item{...}{additional arguments}

\item{refdate}{the curve reference date.}
}
\description{
\code{spotratecurve()} S3 method createas a SpotRateCurve object.
It is dispatched for numeric values, that represent spot rates and
for SpotRate objects.
}
