\name{getSpotRate}
\alias{[[.CurveInterpolation}
\alias{[[.SpotRateCurve}
\alias{getSpotRate}
\title{Get the SpotRate for a given term}
\usage{
  getSpotRate(curve, term)

  \method{[[}{SpotRateCurve} (curve, term)

  \method{[[}{CurveInterpolation} (curve, term)
}
\description{
  Get the SpotRate for a given term
}

