% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{ggplot2-support}
\alias{ggplot2-support}
\alias{autoplot.SpotRateCurve}
\alias{autolayer.SpotRateCurve}
\alias{autolayer.ForwardRate}
\title{ggplot2 plotting functions}
\usage{
\method{autoplot}{SpotRateCurve}(
  object,
  ...,
  curve.name = NULL,
  curve.geom = c("line", "point"),
  curve.interpolation = FALSE,
  curve.x.axis = c("terms", "dates")
)

\method{autolayer}{SpotRateCurve}(
  object,
  ...,
  curve.name = NULL,
  curve.geom = c("line", "point"),
  curve.interpolation = FALSE,
  curve.x.axis = c("terms", "dates")
)

\method{autolayer}{ForwardRate}(
  object,
  ...,
  curve.name = NULL,
  curve.geom = c("step", "line", "point")
)
}
\arguments{
\item{object}{SpotRateCurve or ForwardRate objects}

\item{...}{additional arguments passed to ggplot2 geom_* functions}

\item{curve.name}{Curve's name}

\item{curve.geom}{Curve geom used: \code{line}, \code{point}, \code{step}}

\item{curve.interpolation}{logical indicating to use curve interpolation in
the plot. Defaults to \code{FALSE}.}

\item{curve.x.axis}{x axis can be presented with a numeric scale representing
business days (\code{terms}) or dates (\code{dates}). Defaults to \code{dates}.}
}
\description{
Functions to plot fixedincome objects using ggplot2 package and its grammar
of graphics.
}
