% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hidden.R
\docType{data}
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{Ops,Term,Term-method}
\alias{Arith,SpotRate,SpotRate-method}
\alias{Arith,SpotRate,numeric-method}
\alias{Arith,numeric,SpotRate-method}
\alias{!=,Compounding,Compounding-method}
\alias{==,Compounding,Compounding-method}
\alias{[,Term,numeric,ANY-method}
\alias{[,ForwardRate,missing,missing-method}
\alias{[,ForwardRate,numeric,missing-method}
\alias{[,SpotRateCurve,logical,missing-method}
\alias{[,SpotRateCurve,missing,missing-method}
\alias{[,SpotRateCurve,numeric,missing-method}
\alias{[<-,ForwardRate,numeric,missing,SpotRate-method}
\alias{[<-,SpotRateCurve,logical,missing,numeric-method}
\alias{[<-,SpotRateCurve,numeric,missing,SpotRate-method}
\alias{[[,SpotRateCurve,numeric,missing-method}
\alias{[[,SpotRateCurve,missing,missing-method}
\alias{[[<-,SpotRateCurve,missing,missing,SpotRateCurve-method}
\alias{[[<-,SpotRateCurve,numeric,missing,SpotRate-method}
\alias{[[<-,SpotRateCurve,numeric,missing,numeric-method}
\alias{as.data.frame,ForwardRate-method}
\alias{as.data.frame,SpotRateCurve-method}
\alias{show,Compounding-method}
\alias{show,Daycount-method}
\alias{show,ForwardRate-method}
\alias{show,Interpolation-method}
\alias{show,NelsonSiegel-method}
\alias{show,NelsonSiegelSvensson-method}
\alias{show,SpotRate-method}
\alias{show,SpotRateCurve-method}
\alias{show,Term-method}
\alias{format,Term-method}
\alias{c,Term-method}
\alias{c,SpotRate-method}
\alias{c,SpotRateCurve-method}
\alias{c,ForwardRate-method}
\alias{as.character,Daycount-method}
\alias{as.character,Term-method}
\alias{as.character,SpotRate-method}
\alias{as.character,Compounding-method}
\title{Internal page for hidden aliases}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
