% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotratecurve-interpolation.R
\name{interpolation}
\alias{interpolation}
\alias{interpolation,SpotRateCurve-method}
\alias{interpolation<-,SpotRateCurve,Interpolation-method}
\alias{interpolation<-,SpotRateCurve,NULL-method}
\alias{interpolation<-}
\title{Set/Get interpolation to SpotRateCurve}
\usage{
interpolation(x, ...)

interpolation(x) <- value
}
\arguments{
\item{x}{a SpotRateCurve object.}

\item{...}{additional arguments. Currently unused.}

\item{value}{a Interpolation object.}
}
\value{
A Interpolatin object.
}
\description{
Sets and gets interpolation method to the SpotRateCurve.
}
\examples{
terms <- c(1, 11, 26, 27, 28)
rates <- c(0.0719, 0.056, 0.0674, 0.0687, 0.07)
curve <- spotratecurve(rates, terms, "discrete", "actual/365", "actual")
interpolation(curve) <- interp_flatforward()
interpolation(curve)
}
