% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotratecurve-interpolation.R
\name{prepare_interpolation}
\alias{prepare_interpolation}
\alias{interpolation,SpotRateCurve-method}
\alias{interpolation<-,SpotRateCurve,Interpolation-method}
\alias{interpolation<-,SpotRateCurve,NULL-method}
\title{Create the interpolation function}
\usage{
prepare_interpolation(object, x, ...)
}
\arguments{
\item{x}{a SpotRateCurve object.}

\item{...}{additional arguments. Currently unused.}

\item{value}{a Interpolation object.}
}
\value{
A Interpolatin object.
}
\description{
Creates the interpolation function to a SpotRateCurve object.
}
\examples{
terms <- c(1, 11, 26, 27, 28)
rates <- c(0.0719, 0.056, 0.0674, 0.0687, 0.07)
curve <- spotratecurve(rates, terms, "discrete", "actual/365", "actual")
interpolation(curve) <- interp_flatforward()
interpolation(curve)
}
