\name{rate}
\alias{rate}
\alias{rate.CompoundFactor}
\alias{rate.SpotRate}
\title{rate}
\usage{
rate(object, ...)

\method{rate}{SpotRate}(object)

\method{rate}{CompoundFactor}(object, dib = 252, compounding = "compounded")
}
\value{
\code{interest rate}

\code{implied interest rate}
}
\description{
Interest rate
}
\details{
Interest rate
}

