\name{rates}
\alias{rates}
\alias{rates.CurveInterpolation}
\alias{rates.SpotRateCurve}
\alias{rates.spotrate}
\title{rates}
\usage{
rates(object, ...)

\method{rates}{SpotRateCurve}(object)

\method{rates}{CurveInterpolation}(object, terms = NULL)

\method{rates}{spotrate}(spr)
}
\value{
\code{interest rates}

\code{interest rates}

\code{interest rate}
}
\description{
Interest rates
}
\details{
Interest rates
}

