% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shift}
\alias{shift}
\alias{shift,Term-method}
\alias{shift,numeric-method}
\title{Shift vectors}
\usage{
shift(x, k = 1, ..., fill = NA)
}
\arguments{
\item{x}{a vector object.}

\item{k}{a numeric with the number of elements to shift the Term vector}

\item{...}{additional arguments. Currently unused.}

\item{fill}{a numeric value (or \code{NA}) to fill the empty created by
shifting a vector object.}
}
\value{
A shifted vector object of the same type of provided object.
}
\description{
Element wise shift of vectors by \code{k} positions.
}
\examples{
shift(1:10, fill = 0)

t <- term(1:10, "months")
shift(t)
}
