% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotrate-class.R
\name{spotrate-arith-method}
\alias{spotrate-arith-method}
\alias{Arith,SpotRate,SpotRate-method}
\alias{Arith,SpotRate,numeric-method}
\alias{Arith,numeric,SpotRate-method}
\title{SpotRate arithmetic operations}
\usage{
\S4method{Arith}{SpotRate,SpotRate}(e1, e2)

\S4method{Arith}{SpotRate,numeric}(e1, e2)

\S4method{Arith}{numeric,SpotRate}(e1, e2)
}
\arguments{
\item{e1}{a \code{SpotRate} object or a numeric}

\item{e2}{a \code{SpotRate} object or a numeric

SpotRate objects can be summed among themselves or with numeric variables.

When two SpotRate objects are being summed and their slots are different of
each other, the resulting SpotRate brings the slots of the first object
given (\code{e1}).}
}
\value{
a SpotRate object.
}
\description{
Arithmetic operations with SpotRate class
}
\examples{

spr <- as.spotrate("0.06 simple actual/365 actual")
spr <- spr + 0.01
}
