% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotratecurve-class.R
\docType{class}
\name{SpotRateCurve-class}
\alias{SpotRateCurve-class}
\title{SpotRateCurve class}
\description{
The SpotRateCurve class abstracts a term structure of SpotRate objects.
The SpotRateCurve has a reference date (\code{refdate} slot), that is a
mark to market date.
The SpotRates are indexed to future dates according to its reference date
and these future dates represent the terms of the SpotRateCurve.
}
\details{
Once the SpotRateCurve object is built, any SpotRate can be accessed
using indexing operations: \code{[]} positional indexing, \code{[[]]}
term indexing.

The SpotRateCurve inherits SpotRate class and has three slots:
terms that is a Term object, refdate and interpolation that defines the
method used to interpolate the curve.
}
