% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-class.R
\name{term-class}
\alias{term-class}
\alias{term.numeric}
\alias{term.Date}
\title{Term class}
\usage{
\method{term}{numeric}(x, units = "days")

\method{term}{Date}(x, end_date, calendar)
}
\arguments{
\item{x}{can be a numeric value representing the time period or the initial
date for a period between two dates.}

\item{units}{one of the valid \code{units}: \code{days}, \code{monts},
\code{years}.}

\item{end_date}{the final date for a period between two dates.}

\item{calendar}{the calendar used to compute the amount of days for a period
between two dates.}
}
\description{
It is the time interval used in calculations with interest rates.
The term class represents the period used to discount or compound a spot
rate.
}
\examples{
term(6, "months")
if (require("bizdays")) {
  term(as.Date("2022-02-02"), as.Date("2022-02-23"), "Brazil/ANBIMA")
}
}
