% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-class.R
\name{term-coercion}
\alias{term-coercion}
\alias{as.character,Term-method}
\alias{as.term}
\alias{as.term,character-method}
\title{Term coercion}
\usage{
\S4method{as.character}{Term}(x)

as.term(x, ...)

\S4method{as.term}{character}(x, ...)
}
\arguments{
\item{x}{a Term object or a character representing a term.}
}
\description{
The Term class has a string representation and can be coerced to a character.
Also a character can be coerced to a Term object.
}
\details{
The string representation of the Term class follows the layout:
\code{<number> <units>}
where units is one of: days, months, years.
}
\examples{
t <- as.term("6 months")
as.character(t)

}
