% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-class.R
\name{term-methods}
\alias{term-methods}
\alias{[,Term,ANY,ANY-method}
\alias{c.Term}
\alias{format.Term}
\alias{shift,Term-method}
\alias{diff,Term-method}
\title{Term methods}
\usage{
\S4method{[}{Term,ANY,ANY}(x, i, j, ..., drop = TRUE)

\method{c}{Term}(x, ...)

\method{format}{Term}(x, ...)

\S4method{shift}{Term}(x, k = 1, ..., fill = NA)

\S4method{diff}{Term}(x, lag = 1, ..., fill = NULL)
}
\arguments{
\item{x}{a Term object.}

\item{i}{index used with the \code{[]} operator. Can be numeric
(positional index) or boolean (\code{logical}), similar to vector indexing.}

\item{j}{\emph{unused}.}

\item{...}{additional arguments.}

\item{drop}{\emph{unused}.}

\item{k}{a numeric with the number of elements to shift the Term vector}

\item{fill}{a numeric value (or \code{NA}) to fill the empty created by
applying shift or diff to a Term vector.}

\item{lag}{a numerix indicating which lag to use.}
}
\description{
The Term class has a few methods that helps in tasks involving data
manipulation.
}
\details{
Indexing a Term with the \code{[]} operator is similar to vector indexing.
The only difference is that the returned object is of Term class.

\code{c} concatenates terms and keeps the units correct.

\code{format} coerces the Term class to a string representation that is
compatible with the one used to visualize data.frames.

\code{shift} shifts the Term vector elements by \code{k} positions.

\code{diff} returns a Term vector with lagged differences.
}
\examples{
t <- as.term("6 months")
as.character(t)
}
