% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardrate-class.R
\name{as.forwardrate}
\alias{as.forwardrate}
\alias{as.forwardrate.SpotRate}
\alias{as.forwardrate.SpotRateCurve}
\title{Coerce objects to ForwardRate}
\usage{
as.forwardrate(x, ...)

\method{as.forwardrate}{SpotRate}(x, terms, ...)

\method{as.forwardrate}{SpotRateCurve}(x, ...)
}
\arguments{
\item{x}{a \code{SpotRate} or a \code{SpotRateCurve} object.}

\item{...}{additional arguments}

\item{terms}{a numeric with positive values representing terms or a Term
object.}
}
\value{
A \code{ForwardRate} object created from another object, \code{SpotRate} or
\code{SpotRateCurve}.
}
\description{
A \code{ForwardRate} object can be created from a \code{SpotRate} object and
a \code{SpotRateCurve}.
}
