% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotrate-class.R
\name{as.spotrate}
\alias{as.spotrate}
\alias{as.spotrate,character-method}
\title{Coerce to SpotRate}
\usage{
\S4method{as.spotrate}{character}(x, simplify = TRUE)
}
\arguments{
\item{x}{a character with SpotRate specification.}
}
\description{
Coerce character objects to SpotRate class
}
\details{
The character representation of a SpotRate is as follows:

\preformatted{"RATE COMPOUNDING DAYCOUNT"}

where:
\itemize{
\item \code{RATE} is a numeric value
\item \code{COMPOUNDING} is one of the following:
\code{simple}, \code{discrete}, \code{continuous}
\item \code{DAYCOUNT} is a valid day count rule, pex. \code{business/252},
see \code{\link{daycount-class}}
}
}
\examples{

as.spotrate(c("0.06 simple actual/365", "0.11 discrete business/252"))

}
