\name{calendar}
\alias{calendar}
\alias{calendar.spotrate}
\title{Calendar accessor}
\usage{
calendar(obj, ...)

\method{calendar}{spotrate}(obj, ...)
}
\arguments{
  \item{obj}{any object which has a calendar attribute}

  \item{...}{extra arguments}
}
\value{
a \code{calendar} object
}
\description{
Method to access the \code{calendar} attribute.
}

