% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounding-class.R
\name{compounding-class}
\alias{compounding-class}
\alias{Simple-class}
\alias{Compounding-class}
\alias{Discrete-class}
\alias{Continuous-class}
\alias{compounding}
\title{Compounding class}
\usage{
compounding(x = c("simple", "discrete", "continuous"))
}
\arguments{
\item{x}{a character with the name of compounding regime:
\code{simple}, \code{discrete}, \code{continuous}}
}
\description{
The Compounding class abstracts the compounding regime used to
discount or compound a spot rate.
}
\details{
There are 3 compoundings:
\itemize{
\item \code{simple} for simple interest rate compounding \deqn{1 + rt}
\item \code{discrete} for compounded interest rate compounding
\deqn{(1 + r)^t}
\item \code{continuous} for continuous interest rate compounding
\deqn{exp(rt)}
}

A Compounding object can be instanciated with the \code{compounding}
function, passing a string with the name of one of the compounding regimes:
\code{simple}, \code{discrete}, \code{continuous}.

The \code{compounding} class has 2 methods:
\itemize{
\item \code{compound} to compound the spot rate for a given term.
\item \code{rates} to compute the implied rate for a compound factor
in a given term.
}
}
\examples{
compounding("simple")
compounding("discrete")
compounding("continuous")

comp <- compounding("discrete")
compound(comp, 0.06, 2) # equals (1 + 0.06) ^ 2 = 1.1236
rates(comp, 1.1236, 2) # equals 0.06
}
