\name{compounding-class}
\alias{compound.compounding}
\alias{compounding-class}
\alias{continuousCompounding}
\alias{discreteCompounding}
\alias{rates.compounding}
\alias{simpleCompounding}
\title{Compounding class}
\usage{
discreteCompounding()

simpleCompounding()

continuousCompounding()

\method{compound}{compounding}(comp, value, term)

\method{rates}{compounding}(comp, value, term)
}
\description{
Compounding
}
\details{
Compounding class
}

