% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compounding-class.R
\docType{class}
\name{Compounding-class}
\alias{Compounding-class}
\alias{Simple-class}
\alias{Discrete-class}
\alias{Continuous-class}
\title{Compounding class}
\description{
The Compounding class abstracts the compounding regime used to
discount or compound a spot rate.
}
\details{
There are 3 compoundings:
\itemize{
\item \code{simple} for simple interest rate compounding \deqn{1 + rt}
\item \code{discrete} for compounded interest rate compounding
\deqn{(1 + r)^t}
\item \code{continuous} for continuous interest rate compounding
\deqn{exp(rt)}
}

The \code{Compounding} class has 2 methods:
\itemize{
\item \code{compound} to compound the spot rate for a given term.
\item \code{rates} to compute the implied rate for a compound factor
in a given term.
}
}
