\name{compounding}
\alias{compounding}
\alias{compounding.spotrate}
\title{Compounding accessor}
\usage{
compounding(obj, ...)

\method{compounding}{spotrate}(obj, ...)
}
\arguments{
  \item{obj}{any object which has a compounding attribute}

  \item{...}{extra arguments}
}
\value{
a compounding object
}
\description{
Method to access the compounding attribute.
}

