\docType{data}
\name{Compounding}
\alias{Compounding}
\title{Compounding Functinos}
\format{List of 3
 $ compounded: atomic [1:1] compounded
  ..- attr(*, "compound")=function (value, term, dib)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 8 37 8 86 37 86 8 8
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1050fc1c0> 
  ..- attr(*, "implied.rate")=function (value, term, dib)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 9 41 9 88 41 88 9 9
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1050fc1c0> 
 $ simple    : atomic [1:1] simple
  ..- attr(*, "compound")=function (value, term, dib)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 11 33 11 80 33 80 11 11
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1050fc1c0> 
  ..- attr(*, "implied.rate")=function (value, term, dib)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 12 37 12 86 37 86 12 12
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1050fc1c0> 
 $ continuous: atomic [1:1] continuous
  ..- attr(*, "compound")=function (value, term, dib)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 14 37 14 83 37 83 14 14
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1050fc1c0> 
  ..- attr(*, "implied.rate")=function (value, term, dib)  
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 15 41 15 89 41 89 15 15
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x1050fc1c0>}
\usage{
  Compounding
}
\description{
  Compounding Functinos
}
\keyword{datasets}

