% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daycount-class.R
\name{daycount-class}
\alias{daycount-class}
\alias{daycount}
\alias{dib,Daycount-method}
\alias{toyears,Daycount,Term,missing-method}
\alias{toyears,Daycount,character,missing-method}
\alias{toyears,Daycount,numeric,character-method}
\title{Daycount class}
\usage{
daycount(x, ...)

\S4method{dib}{Daycount}(x)

\S4method{toyears}{Daycount,Term,missing}(x, t, units)

\S4method{toyears}{Daycount,character,missing}(x, t, units)

\S4method{toyears}{Daycount,numeric,character}(x, t, units)
}
\arguments{
\item{x}{a character representing a daycount rule, like: \code{business/252},
\code{actual/365}, \code{actual/360}, ...}

\item{t}{represents the term to compound. Can be a numeric, a \code{Term},
or a character representing a \code{Term}. See Details.}

\item{units}{a character with the Term units. Can also be missing.
See Details.}
}
\description{
\code{Daycount} class helps computing terms used to compound interest rates.
}
\details{
The daycount convention determines the amount of days, in years, represents
the given term.

\code{daycount} creates a \code{Daycount} object and accepts the
following daycount rules: \code{actual/365}, \code{actual/360},
\code{business/252}

The method \code{dib} returns the days in base for a daycount convention.
Since we work with annual rates the days in base define the amount of days
in a year used in the convention.

\code{toyears} returns the given term in years, since we are assuming
annual rates.
The \code{t} argument can be a term instance, a string defining a term
or a numeric.
In the last alternative, the \code{units} argument must be
provided with a valid Term units (days, months, years).
}
\examples{
dc <- daycount("actual/360")
dib(dc)
toyears(dc, 10, "days")
}
