% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daycount-class.R
\name{daycount}
\alias{daycount}
\title{Create Daycount class}
\usage{
daycount(x, ...)
}
\arguments{
\item{x}{a character representing a daycount rule, like: \code{business/252},
\code{actual/365}, \code{actual/360}, ...}

\item{...}{additional arguments. Currently unused.}
}
\value{
A \code{Daycount} object.
}
\description{
\code{daycount} creates a \code{Daycount} object.
It accepts the following daycount rules: \code{actual/365},
\code{actual/360}, \code{business/252}.
}
\examples{
dc <- daycount("actual/360")
}
