% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedincome-internal.R
\docType{package}
\name{fixedincome-package}
\alias{fixedincome-package}
\alias{fixedincome}
\title{Fixed income models, calculations and data structures}
\description{
The \code{fixedincome} package brings a set of funtions that helps with the
mathematics of interest rates and fixed income.
It handles the interest rates and term structures of interest rates as objects and
provides many methods to tackle specific issues
like computing discount factors and forward rates, interpolate term structures,
fit curve models and so much more.
This package also supports methods and models commonly used by practitioners to
do fixed income calculations.
}
\references{
Frank Fabozzi. Fixed Income Mathematics, Wiley, 1994.

Bruce Tuckman. Fixed Income Securities, Wiley, 1994.
}
\author{
Wilson Freitas \email{wilson.freitas@gmail.com}
}
