% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixedincome-internal.R
\docType{package}
\name{fixedincome-package}
\alias{fixedincome-package}
\alias{fixedincome}
\title{Fixed income models, calculations, data structures and instruments}
\description{
\code{fixedincome} has a set of funtions which helps with the
mathematics of interest rates and fixed income.
It handles the interest rates and compounding
factors as objects and provides many methods to tackle specific issues
like compute discount factors, find equivalent rates, forward rates,
and so on. It also has classes to represent commom fixed income entities such as
a term structure of interest rates, spot rates and day count rules.
This package also supports methods and models commom used by practitioners to
do fixed income calculations.
}
\references{
Frank Fabozzi. Fixed Income Mathematics, Wiley, 1994.

Bruce Tuckman. Fixed Income Securities, Wiley, 1994.
}
\author{
Wilson Freitas \email{wilson.freitas@gmail.com}
}
