% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardrate-class.R
\docType{class}
\name{ForwardRate-class}
\alias{ForwardRate-class}
\title{ForwardRate class}
\description{
ForwardRate class abstracts a forward rate and differs from a spot rate
because it has a term associated.
}
