% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forwardrate-class.R
\name{forwardrate-method}
\alias{forwardrate-method}
\alias{forwardrate}
\alias{forwardrate.numeric}
\alias{forwardrate.SpotRateCurve}
\title{Create a ForwardRate object}
\usage{
\method{forwardrate}{numeric}(x, terms, compounding, daycount, calendar, .copyfrom = NULL, ...)

\method{forwardrate}{SpotRateCurve}(x, t1 = NULL, t2 = NULL, ...)
}
\arguments{
\item{x}{a numeric or a SpotRateCurve object.}

\item{terms}{a numeric vector with positive values representing terms of
the forward rates.}

\item{compounding}{a character with the compouning name.}

\item{daycount}{a character representing the daycount.}

\item{calendar}{a calendar object.}

\item{.copyfrom}{a SpotRate object that is used as reference to build
the SpotRateCurve object.}

\item{...}{additional arguments.

The arguments \code{t1} and \code{t2} define initial and final term used to
extract a ForwardRate from a SpotRateCurve.}

\item{t1}{initial term}

\item{t2}{final term}
}
\description{
\code{forwardrate()} S3 method creates a ForwardRate object.
}
