\name{generic-SpotRateCurve}
\alias{as.data.frame.SpotRateCurve}
\alias{generic-SpotRateCurve}
\alias{head.SpotRateCurve}
\alias{length.SpotRateCurve}
\alias{plot.SpotRateCurve}
\alias{tail.SpotRateCurve}
\title{SpotRateCurve generic extensions}
\usage{
  \method{as.data.frame}{SpotRateCurve} (curve, ...)

  \method{length}{SpotRateCurve} (curve)

  \method{plot}{SpotRateCurve} (curve, ...)

  \method{head}{SpotRateCurve} (curve, n = 6L, ...)

  \method{tail}{SpotRateCurve} (curve, n = 6L, ...)
}
\description{
  SpotRateCurve generic extensions
}

