\name{[[.CurveInterpolation}
\alias{[[.CurveInterpolation}
\alias{[[.SpotRateCurve}
\alias{getSpotRate}
\title{Get the SpotRate for a given term}
\usage{
\method{[[}{CurveInterpolation}(curve, term)

getSpotRate(curve, term)

\method{[[}{SpotRateCurve}(curve, term)
}
\description{
Get the SpotRate for a given term
}

