% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation-class.R
\docType{class}
\name{Interpolation-class}
\alias{Interpolation-class}
\alias{FlatForward-class}
\alias{Linear-class}
\alias{LogLinear-class}
\alias{NaturalSpline-class}
\alias{HermiteSpline-class}
\alias{MonotoneSpline-class}
\alias{NelsonSiegel-class}
\alias{NelsonSiegelSvensson-class}
\title{Interpolation classes}
\description{
Classes that implement interpolation methods to be used with SpotRateCurve
objects.
Every class that implement a interpolation method inherits the
\code{Interpolation} class.
}
