\name{rates}
\alias{rates}
\alias{rates.CurveInterpolation}
\alias{rates.SpotRateCurve}
\title{rates}
\usage{
rates(object, ...)

\method{rates}{SpotRateCurve}(object)

\method{rates}{CurveInterpolation}(object, terms = NULL)
}
\value{
\code{interest rates}

\code{interest rates}
}
\description{
Interest rates
}
\details{
Interest rates
}

