\name{rates}
\alias{rates}
\alias{rates.spotrate}
\title{Compound function}
\usage{
rates(obj, ...)

\method{rates}{spotrate}(obj, ...)
}
\arguments{
  \item{obj}{.}

  \item{...}{.}
}
\description{
Computes the compounding factor
}
\details{
Compounding factor

The \code{rates.spotrate} returns the numeric value
representing the spot rates.
}

