% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotrate-class.R
\docType{class}
\name{SpotRate-class}
\alias{SpotRate-class}
\title{SpotRate class}
\description{
The \code{SpotRate} class abstracts the interst rate and has methods
to handle many calculations on it.
}
\details{
The \code{SpotRate} class fully specifies spot rates.
It has:
\itemize{
\item the spot rate values which are numeric values representing the rate.
\item the compounding regime that specifies how to compound the spot
rate. This is a \code{Compounding} object.
\item the daycount rule to compute the compounding periods right
adjusted to the spot rate frequency.
\item the calendar according to which the number of days are counted.
}

The \code{SpotRate} class is a \code{numeric}, that represents the
interest rate and that has the slots: \code{compounding}, \code{daycount}
and \code{calendar}.

For example, an annual simple interest rate of 6\%, that compounds in
calendar days, is defined as follows:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{sr_simple <- spotrate(0.06, "simple", "actual/360", "actual")
sr_simple
#> [1] "0.06 simple actual/360 actual"
}\if{html}{\out{</div>}}

\code{actual/360} is the daycount rule and \code{actual} is the calendar.

Differently, an annual compound interest rate of 10\%, that compounds
in business days according to calendar \code{Brazil/ANBIMA} is

\if{html}{\out{<div class="sourceCode r">}}\preformatted{sr_disc <- spotrate(0.1, "discrete", "business/252", "Brazil/ANBIMA")
sr_disc
#> [1] "0.1 discrete business/252 Brazil/ANBIMA"
}\if{html}{\out{</div>}}

The \code{calendar} slot is a \code{bizdays} calendar.

An $100,000 investment in an instrument that pays that interst rate for 5
years has the future value.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{100000 * compound(sr_disc, term(5, "years"))
#> [1] 161051
}\if{html}{\out{</div>}}

for the simple interest rate we have

\if{html}{\out{<div class="sourceCode r">}}\preformatted{100000 * compound(sr_simple, term(5, "years"))
#> [1] 130000
}\if{html}{\out{</div>}}

SpotRate objects can be created with vectors

\if{html}{\out{<div class="sourceCode r">}}\preformatted{sr_vec <- spotrate(abs(rnorm(10)), "discrete", "business/252", "Brazil/ANBIMA")
sr_vec
#>  [1] "0.86672532 discrete business/252 Brazil/ANBIMA"
#>  [2] "0.03464431 discrete business/252 Brazil/ANBIMA"
#>  [3] "0.85367021 discrete business/252 Brazil/ANBIMA"
#>  [4] "0.34312519 discrete business/252 Brazil/ANBIMA"
#>  [5] "0.87847078 discrete business/252 Brazil/ANBIMA"
#>  [6] "0.94804384 discrete business/252 Brazil/ANBIMA"
#>  [7] "0.30661161 discrete business/252 Brazil/ANBIMA"
#>  [8] "0.53746114 discrete business/252 Brazil/ANBIMA"
#>  [9] "1.45423675 discrete business/252 Brazil/ANBIMA"
#> [10] "0.07353385 discrete business/252 Brazil/ANBIMA"
}\if{html}{\out{</div>}}

and can be put into a \code{data.frame}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data.frame(spot_rate = sr_vec)
#>                                         spot_rate
#> 1  0.86672532 discrete business/252 Brazil/ANBIMA
#> 2  0.03464431 discrete business/252 Brazil/ANBIMA
#> 3  0.85367021 discrete business/252 Brazil/ANBIMA
#> 4  0.34312519 discrete business/252 Brazil/ANBIMA
#> 5  0.87847078 discrete business/252 Brazil/ANBIMA
#> 6  0.94804384 discrete business/252 Brazil/ANBIMA
#> 7  0.30661161 discrete business/252 Brazil/ANBIMA
#> 8  0.53746114 discrete business/252 Brazil/ANBIMA
#> 9  1.45423675 discrete business/252 Brazil/ANBIMA
#> 10 0.07353385 discrete business/252 Brazil/ANBIMA
}\if{html}{\out{</div>}}

once in a \code{data.frame}, dplyr verbs can be used to manipulate it.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{require(dplyr, warn.conflicts = FALSE)
#> Loading required package: dplyr

data.frame(spot_rate = sr_vec) |>
   mutate(comp = compound(spot_rate, term(5, "months")))
#>                                         spot_rate     comp
#> 1  0.86672532 discrete business/252 Brazil/ANBIMA 1.297030
#> 2  0.03464431 discrete business/252 Brazil/ANBIMA 1.014292
#> 3  0.85367021 discrete business/252 Brazil/ANBIMA 1.293243
#> 4  0.34312519 discrete business/252 Brazil/ANBIMA 1.130790
#> 5  0.87847078 discrete business/252 Brazil/ANBIMA 1.300425
#> 6  0.94804384 discrete business/252 Brazil/ANBIMA 1.320280
#> 7  0.30661161 discrete business/252 Brazil/ANBIMA 1.117878
#> 8  0.53746114 discrete business/252 Brazil/ANBIMA 1.196286
#> 9  1.45423675 discrete business/252 Brazil/ANBIMA 1.453668
#> 10 0.07353385 discrete business/252 Brazil/ANBIMA 1.030006
}\if{html}{\out{</div>}}

SpotRate is \code{numeric}, so it executes arithmetic and comparison operations
with \code{numeric} objects.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{data.frame(spot_rate = sr_vec) |>
   mutate(
     new_spot_rate = spot_rate + 0.02,
     check_gt_1pp = spot_rate > 0.01,
     check_gt_nsr = spot_rate > new_spot_rate
   )
#>                                         spot_rate
#> 1  0.86672532 discrete business/252 Brazil/ANBIMA
#> 2  0.03464431 discrete business/252 Brazil/ANBIMA
#> 3  0.85367021 discrete business/252 Brazil/ANBIMA
#> 4  0.34312519 discrete business/252 Brazil/ANBIMA
#> 5  0.87847078 discrete business/252 Brazil/ANBIMA
#> 6  0.94804384 discrete business/252 Brazil/ANBIMA
#> 7  0.30661161 discrete business/252 Brazil/ANBIMA
#> 8  0.53746114 discrete business/252 Brazil/ANBIMA
#> 9  1.45423675 discrete business/252 Brazil/ANBIMA
#> 10 0.07353385 discrete business/252 Brazil/ANBIMA
#>                                     new_spot_rate check_gt_1pp check_gt_nsr
#> 1  0.88672532 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
#> 2  0.05464431 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
#> 3  0.87367021 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
#> 4  0.36312519 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
#> 5  0.89847078 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
#> 6  0.96804384 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
#> 7  0.32661161 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
#> 8  0.55746114 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
#> 9  1.47423675 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
#> 10 0.09353385 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
}\if{html}{\out{</div>}}

SpotRate vectors also are created with the concatenation function \code{c}.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{c(sr_disc, 0.1, 0.13, 0.14, 0.15)
#> [1] "0.10 discrete business/252 Brazil/ANBIMA"
#> [2] "0.10 discrete business/252 Brazil/ANBIMA"
#> [3] "0.13 discrete business/252 Brazil/ANBIMA"
#> [4] "0.14 discrete business/252 Brazil/ANBIMA"
#> [5] "0.15 discrete business/252 Brazil/ANBIMA"
}\if{html}{\out{</div>}}

Furtherly, all indexing operations of numeric objects are supported by
SpotRate objects.
\subsection{Invalid Operations}{

Operations involving SpotRate objects with different \code{compounding},
\code{daycount} or \code{calendar}, raise errors.

This happens with the following operations:
\itemize{
\item Compare: >, <, <=, >=
\item Arithmetic: +, -, *, /
\item Concatenation: \code{c}
}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{try(sr_simple + sr_disc)
#> Error in stop_if_spotrate_slots_differ(e1, e2, "SpotRate objects have different slots") : 
#>   SpotRate objects have different slots
try(sr_simple > sr_disc)
#> Error in stop_if_spotrate_slots_differ(e1, e2, "SpotRate objects have different slots") : 
#>   SpotRate objects have different slots
try(c(sr_simple, sr_disc))
#> Error in stop_if_spotrate_slots_differ(x, values_, "SpotRate objects have different slots") : 
#>   SpotRate objects have different slots
}\if{html}{\out{</div>}}
}
}
\note{
The \code{SpotRate} objects are annual rates.
}
