% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotrate-class.R
\name{spotrate-class}
\alias{spotrate-class}
\alias{SpotRate-class}
\alias{spotrate}
\title{SpotRate class}
\usage{
spotrate(x, compounding, daycount, calendar, .copyfrom = NULL)
}
\arguments{
\item{x}{a numeric vector representing spot rate values.}

\item{compounding}{a \code{Compounding} object.}

\item{daycount}{a \code{Daycount} object.}

\item{calendar}{a \code{bizdays} calendar.}

\item{.copyfrom}{a \code{SpotRate} object used as reference to copy
attributes.}
}
\description{
The \code{SpotRate} class abstracts the interst rate and has methods
to handle many calculations on it.
}
\details{
The \code{SpotRate} class fully specifies spot rates.
It has:
\itemize{
\item the spot rate values which are numeric values representing the rate.
\item the compounding regime that specifies how to compound the spot
rate. This is a \code{Compounding} object.
\item the daycount rule to compute the compounding periods right
adjusted to the spot rate frequency (which is annual).
\item the calendar according to which the number of days are counted.
}

The \code{SpotRate} class is a \code{numeric} and
all values in it share the \code{compounding}, \code{daycount}
and \code{calendar} slots.

The \code{calendar} attribute is an instance of \code{bizdays}
\code{Calendar} class or the name of one of the calendars
that are already registered in bizdays register.
}
\note{
The \code{SpotRate} objects are annual rates.
}
\examples{
spotrate(0.06, "continuous", "actual/365", "actual")
spotrate(c(0.06, 0.07, 0.08), "continuous", "actual/365", "actual")
}
