% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotrate-class.R
\name{spotrate-compare-method}
\alias{spotrate-compare-method}
\alias{Compare,SpotRate,SpotRate-method}
\alias{Compare,SpotRate,numeric-method}
\alias{Compare,numeric,SpotRate-method}
\title{SpotRate comparison operations}
\usage{
\S4method{Compare}{SpotRate,SpotRate}(e1, e2)

\S4method{Compare}{SpotRate,numeric}(e1, e2)

\S4method{Compare}{numeric,SpotRate}(e1, e2)
}
\arguments{
\item{e1}{a \code{SpotRate} object or a numeric}

\item{e2}{a \code{SpotRate} object or a numeric}
}
\value{
a boolean \code{logical} object.
}
\description{
Comparison operations with SpotRate class
\code{SpotRate} objects can be compared among themselves or with numeric
variables.
}
\examples{

spr <- as.spotrate("0.06 simple actual/365 actual")
spr == 0.06
}
