\name{SpotRate}
\alias{SpotRate}
\title{SpotRate - it is the pair (InterestRate, term)
A SpotRateCurve is a sorted set of SpotRates}
\usage{
  SpotRate(value, term, dib = 252,
    compounding = "compounded")
}
\description{
  SpotRate - it is the pair (InterestRate, term) A
  SpotRateCurve is a sorted set of SpotRates
}

