% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotrate-class.R
\name{spotrate}
\alias{spotrate}
\title{Create SpotRate objects}
\usage{
spotrate(x, compounding, daycount, calendar, .copyfrom = NULL)
}
\arguments{
\item{x}{a numeric vector representing spot rate values.}

\item{compounding}{a \code{Compounding} object.}

\item{daycount}{a \code{Daycount} object.}

\item{calendar}{a \code{bizdays} calendar.}

\item{.copyfrom}{a \code{SpotRate} object used as reference to copy
attributes.}
}
\description{
\code{spotrate()} function creates \code{SpotRate} objects.
}
\examples{
spotrate(0.06, "continuous", "actual/365", "actual")
spotrate(c(0.06, 0.07, 0.08), "continuous", "actual/365", "actual")
}
