% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-class.R
\docType{class}
\name{Term-class}
\alias{Term-class}
\alias{DateRangeTerm-class}
\title{Term class}
\description{
It is the time interval used in calculations with interest rates.
The term class represents the period used to discount or compound a spot
rate.
It can be Term object or a DateRangeTerm which defines start and end dates
and a calendar to count the amount of working days between these two dates.
}
